Use color-schema independent variable names for UI
Before this commit we just use silver as foreground and black as
background for some UI elements, however this is not suitable for light
theme. Instead of renaming all silver to black, this commit just use
color-schema independent variable names for them so we can easily sync
code between dark theme and light theme.